# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LG R290 Heat Pump API
  description: "\n# LG R290 Heat Pump Control API\n\nREST API for monitoring and controlling\
    \ LG R290 7kW heat pump via Modbus TCP.\n\n## Features\n- **Heat Pump Control**:\
    \ Power ON/OFF, temperature setpoint control\n- **AI Mode**: Automatic flow temperature\
    \ adjustment based on outdoor temperature and heating curves\n- **Scheduler**:\
    \ Time-based automatic temperature control (weekday/weekend patterns)\n- **Real-time\
    \ Monitoring**: Live status updates via polling\n\n## Modes\n- **Manual Mode**:\
    \ Direct control of flow temperature setpoint\n- **AI Mode (Automatic)**: System\
    \ calculates optimal flow temperature based on:\n  - Outdoor temperature (from\
    \ Shelly BT sensor)\n  - Target room temperature (from thermostat)\n  - Selected\
    \ heating curve (Comfort/Standard/Eco)\n\n## Architecture\n- **Service**: FastAPI\
    \ backend (this API)\n- **Heat Pump**: LG R290 via Modbus TCP\n- **Thermostat**:\
    \ Shelly BT temperature control system\n- **UI**: Web interface for control and\
    \ monitoring\n    "
  version: 1.0.0
paths:
...
  /power:
    post:
      tags:
      - Heat Pump
      summary: Control Heat Pump Power
      description: "Turn the heat pump ON or OFF.\n\n    **Note**: In AI Mode, the\
        \ system may automatically turn the heat pump ON/OFF\n    based on outdoor\
        \ temperature and heating curve calculations.\n\n    **Effect**: Writes to\
        \ Modbus coil register 1 (Enable/Disable Heating/Cooling)."
      operationId: set_power_power_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PowerControl'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /setpoint:
    post:
      tags:
      - Heat Pump
      summary: Set Flow Temperature Setpoint
      description: "Set target flow temperature (water outlet temperature) in \xB0\
        C.\n\n    **Valid Range**: 20.0\xB0C - 60.0\xB0C\n\n    **Manual Mode**: Temperature\
        \ is set directly and maintained.\n\n    **AI Mode**: This setting will be\
        \ overridden by automatic calculations based on:\n    - Outdoor temperature\n\
        \    - Target room temperature\n    - Selected heating curve\n\n    **Effect**:\
        \ Writes to Modbus holding register 40003 (Target Temperature Circuit 1)."
      operationId: set_temperature_setpoint_setpoint_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemperatureSetpoint'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /ai-mode:
    get:
      tags:
      - AI Mode
      summary: Get AI Mode Status
      description: "Get current AI Mode status and detailed information including:\n\
        \    - Enabled/disabled state\n    - Last update timestamp\n    - Current\
        \ outdoor and target room temperatures\n    - Calculated flow temperature\n\
        \    - Selected heating curve\n    - Update interval and adjustment threshold\n\
        \n    **Response includes**:\n    - `enabled`: AI Mode state\n    - `outdoor_temperature`:\
        \ Current outdoor temp (\xB0C)\n    - `target_room_temperature`: Target room\
        \ temp from thermostat (\xB0C)\n    - `calculated_flow_temperature`: Optimal\
        \ flow temp (\xB0C)\n    - `heating_curve`: Selected curve info (name, target\
        \ temp range)"
      operationId: get_ai_mode_ai_mode_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - AI Mode
      summary: Enable/Disable AI Mode
      description: "Enable or disable AI Mode (automatic flow temperature control).\n\
        \n    **AI Mode Enabled (Automatic)**:\n    - System automatically calculates\
        \ optimal flow temperature\n    - Based on outdoor temperature and target\
        \ room temperature\n    - Uses predefined heating curves (Comfort/Standard/Eco)\n\
        \    - Adjusts every 30 seconds\n    - May turn heat pump ON/OFF based on\
        \ outdoor cutoff temperature\n\n    **AI Mode Disabled (Manual)**:\n    -\
        \ Flow temperature is controlled manually via /setpoint\n    - No automatic\
        \ adjustments\n\n    **Default**: AI Mode is enabled by default at startup\
        \ for deterministic control."
      operationId: set_ai_mode_ai_mode_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIModeControl'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AIModeControl:
      properties:
        enabled:
          type: boolean
          title: Enabled
          description: Set to true to enable AI Mode (automatic), false for manual
            control
      type: object
      required:
      - enabled
      title: AIModeControl
      description: Enable or disable AI Mode (automatic flow temperature control based
        on heating curves).
      examples:
      - enabled: true
      - enabled: false
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PowerControl:
      properties:
        power_on:
          type: boolean
          title: Power On
          description: Set to true to turn ON, false to turn OFF
      type: object
      required:
      - power_on
      title: PowerControl
      description: Control heat pump power state.
      examples:
      - power_on: true
      - power_on: false
    TemperatureSetpoint:
      properties:
        temperature:
          type: number
          maximum: 60.0
          minimum: 20.0
          title: Temperature
          description: "Target flow temperature in \xB0C"
          examples:
          - 35.0
          - 40.0
          - 45.0
      type: object
      required:
      - temperature
      title: TemperatureSetpoint
      description: Set target flow temperature (water outlet temperature).
      examples:
      - temperature: 35.0
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError